updater: get blinky test working
This blinky test tries to blink the two LEDs at the same time.  One LED
is timed using the RTC, and the other is timed using the main CPU.  If
the CPU is running at 21 MHz (it is), they should blink at approximately
the same rate.

There is some variance, but mostly they match.

Signed-off-by: Sean Cross <sean@xobs.io>