toboot: fix pre-V2 flash clearing
The `tb_valid_signature_at_page()` function returns negative if the
specified page does not have a valid signature.  The detection of this
logic caused the function to accidentally clear all pages that were NOT
valid signatures, having the opposite effect.

This was combined with a second bug that caused us to clear the pages
belonging to Toboot itself, due to not properly checking
tb_state.clear_hi.

This patch corrects both issues.

Signed-off-by: Sean Cross <sean@xobs.io>